+++ modules/gui-core-bukkit/src/main/kotlin/ru/astrainteractive/astrarating/gui/util/ConfigExt.kt
@@ -1,22 +1,8 @@
-@file:Suppress("Filename")
-
 package ru.astrainteractive.astrarating.gui.util
 
 import org.bukkit.Material
 import org.bukkit.inventory.ItemStack
 import ru.astrainteractive.astrarating.core.settings.AstraRatingConfig
-import kotlin.math.max
-
-internal fun subListFromString(text: String, threshold: Int, cutWords: Boolean): List<String> {
-    val res = if (cutWords) {
-        text.chunked(threshold)
-    } else {
-        text.split(" ").chunked(max(1, (text.length) / threshold)).map {
-            it.joinToString(" ")
-        }
-    }
-    return res
-}
 
 internal fun AstraRatingConfig.Gui.Buttons.Button.toItemStack(): ItemStack {
     val material = Material.getMaterial(material.uppercase()) ?: Material.PAPER
